Begin `list` command work
Thought it'd be helpful to get a specific list of a single facet I want
to focus on. For example, maybe I'm feeling reflective and want to see a
list of things that used to worry over time. So I use `feel list -c` for
all the "concerned" entries I wrote.